Update recipient message status as we fetch the batch
We need to set a unique message reference for each recipient in our batch. The message reference will allow us to identify the status of each message in our batch.
We also need to set a default message status of 'requested' and later a status of 'sent' once we have successfully sent the batch to the communication management API.
This commit refactors the logic around getting recipients, assigning message references and message statuses and updating Oracle.