Perform updates to status and id atomically
When we fetch the next batch of recipients, the message status is automatically set to 'REQUESTED' by the underlying stored procedure f_get_next_batch.
This means that we don't need to set this status alongside the message_id.
We still need to set a message_status of SENT or SENDING (tbc) when the
API call succeeds.