Seed environment variables from secrets in the lambda handler
Attempting to do this at module level creates a race condition where the secretsmanager endpoint is not ready at the time we try to read from secretsmanager.
Instead populate the environment on the first lambda handler function call and set a environment variable to denote this so we don't repeat on every handler call.