Don't raise when no routing plan id is returned
We need to repeatedly call the database for the next batch of recipients, we will do this until no routing plan id is returned.
We don't want to raise when no id is returned.
This commit also moves from OOP style to functional as the class itself held very little instance data.